refactor: crew 상세 페이지 컴포넌트 폴더 구조 변경
_components 폴더 내부로 이동
